# api/openapi.yaml
openapi: 3.1.0
info:
  title: Cubit Control Plane API
  version: 1.0.0
  description: >-
    Strict REST API for Cubit, an open-source bare-metal PaaS for running
    Cloudflare Workers and Durable Objects using celld, Traefik, and S3 storage.
servers:
  - url: /api/v1
    description: Cubit API v1
paths:
...
  /applications/{id}/deploy/direct:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
    post:
      summary: Direct bundle deployment (e.g. from wrangler CLI or CI/CD)
      operationId: deployDirect
      tags:
        - deployments
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DeployDirectRequest'
          multipart/form-data:
            schema:
              type: object
              properties:
                bundle:
                  type: string
                  format: binary
                commitMessage:
                  type: string
                wranglerConfig:
                  type: string
      responses:
        '201':
          description: Direct deployment deployed and active
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Deployment'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
...
components:
  responses:
    BadRequest:
      description: Bad request / validation failure
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
...
    DeploymentStatus:
      type: string
      enum:
        - pending
        - building
        - deploying
        - active
        - failed
        - superseded
    Deployment:
      type: object
      required:
        - id
        - applicationId
        - commitHash
        - status
        - createdAt
      properties:
        id:
          type: string
          format: uuid
        applicationId:
          type: string
          format: uuid
        buildVersion:
          type: integer
          example: 1
          description: "Sequential build version for this application (e.g. 1 for v1)"
        commitHash:
          type: string
          example: "7f8b92c"
        commitMessage:
          type: string
          example: "feat: Add new payment endpoint"
        status:
          $ref: '#/components/schemas/DeploymentStatus'
        bundleSize:
          type: integer
          example: 452000
        errorMessage:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        finishedAt:
          type: string
          format: date-time
          nullable: true
...
    DeployDirectRequest:
      type: object
      required:
        - bundle
      properties:
        bundle:
          type: string
          description: JavaScript bundle code
        commitMessage:
          type: string
          description: Optional deployment message or commit description
        wranglerConfig:
          type: string
          description: Optional wrangler.json or wrangler.toml configuration content
...
    ErrorResponse:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          example: "NOT_FOUND"
        message:
          type: string
          example: "Application with ID not found"
        details:
          type: array
          items:
            type: string
          nullable: true